Corrected wrong path for bh.om.dll reference. Files and folder tidy up for compliance. Renamed Replace() method to CRUD() for compliance
