+++ S_Frame_Adapter/CRUD/Delete/_Delete.cs
