+++ S_Frame_Adapter/CRUD/Read/Bar.cs
