+++ S_Frame_Adapter/CRUD/Read/SectionProperty.cs
