+++ S_Frame_Adapter/CRUD/Read/_Read.cs
