+++ S_Frame_Adapter/S_FrameAdapter.cs
@@ -42,50 +42,7 @@ namespace BH.Adapter.S_Frame
 
         private string m_FolderPath = ("C:/Users/jtaylor/OneDrive - BuroHappold/Tools in Progress/SConcrete_Toolkit");
 
-        /***************************************************/
-        /**** Overrides                                 ****/
-        /***************************************************/
-
-        public override List<IObject> Push(IEnumerable<IObject> objects, string tag = "", Dictionary<string, object> config = null)
-        {
-            bool success = true;
-
-            string pushType;
-
-            object ptObj;
-            if (config != null && config.TryGetValue("PushType", out ptObj))
-                pushType = ptObj.ToString();
-            else
-                pushType = "Replace";
-
-            List<IObject> objectsToPush = Config.CloneBeforePush ? objects.Select(x => x is BHoMObject ? ((BHoMObject)x).GetShallowClone() : x).ToList() : objects.ToList(); //ToList() necessary for the return collection to function properly for cloned objects
-
-            Type iBHoMObjectType = typeof(IBHoMObject);
-            Type iResultType = typeof(IResult);
-            MethodInfo miToList = typeof(Enumerable).GetMethod("Cast");
-            foreach (var typeGroup in objectsToPush.GroupBy(x => x.GetType()))
-            {
-                MethodInfo miListObject = miToList.MakeGenericMethod(new[] { typeGroup.Key });
-
-                var list = miListObject.Invoke(typeGroup, new object[] { typeGroup });
-
-                if (iBHoMObjectType.IsAssignableFrom(typeGroup.Key))
-                {
-                    if (pushType == "Replace")
-                        success &= Replace(list as dynamic, tag);
-                    else if (pushType == "UpdateOnly")
-                    {
-                        success &= UpdateOnly(list as dynamic, tag);
-                    }
-                }
-                else if (iResultType.IsAssignableFrom(typeGroup.Key))
-                {
-                    success &= Create(list as dynamic);
-                }
-            }
-
-            return success ? objectsToPush : new List<IObject>();
-        }
+  
 
         /***************************************************/
     }
