+++ S_Frame_Adapter/S_Frame_Adapter.csproj
@@ -34,9 +34,9 @@
       <SpecificVersion>False</SpecificVersion>
       <HintPath>..\..\BHoM\Build\Analytical_oM.dll</HintPath>
     </Reference>
-    <Reference Include="BHoM">
+    <Reference Include="BHoM, Version=3.0.0.0, Culture=neutral, processorArchitecture=MSIL">
       <SpecificVersion>False</SpecificVersion>
-      <HintPath>..\..\..\..\BHoM\Build\BHoM.dll</HintPath>
+      <HintPath>..\..\BHoM\Build\BHoM.dll</HintPath>
     </Reference>
     <Reference Include="BHoM_Adapter">
       <HintPath>..\..\BHoM_Adapter\Build\BHoM_Adapter.dll</HintPath>
@@ -91,20 +91,21 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Create\BarResult.cs" />
-    <Compile Include="Create\Bar.cs" />
-    <Compile Include="Create\Material.cs" />
-    <Compile Include="Create\Node.cs" />
-    <Compile Include="Create\SectionProperty.cs" />
-    <Compile Include="Create\_Create.cs" />
-    <Compile Include="Delete\_Delete.cs" />
+    <Compile Include="AdapterActions\Push.cs" />
+    <Compile Include="CRUD\Create\BarResult.cs" />
+    <Compile Include="CRUD\Create\Bar.cs" />
+    <Compile Include="CRUD\Create\Material.cs" />
+    <Compile Include="CRUD\Create\Node.cs" />
+    <Compile Include="CRUD\Create\SectionProperty.cs" />
+    <Compile Include="CRUD\Create\_Create.cs" />
+    <Compile Include="CRUD\Delete\_Delete.cs" />
     <Compile Include="Types\NextId.cs" />
-    <Compile Include="Read\Bar.cs" />
-    <Compile Include="Read\Material.cs" />
-    <Compile Include="Read\Node.cs" />
-    <Compile Include="Read\SectionProperty.cs" />
-    <Compile Include="Read\_Read.cs" />
-    <Compile Include="Update\_UpdateObjects.cs" />
+    <Compile Include="CRUD\Read\Bar.cs" />
+    <Compile Include="CRUD\Read\Material.cs" />
+    <Compile Include="CRUD\Read\Node.cs" />
+    <Compile Include="CRUD\Read\SectionProperty.cs" />
+    <Compile Include="CRUD\Read\_Read.cs" />
+    <Compile Include="CRUD\Update\_UpdateObjects.cs" />
     <Compile Include="S_FrameAdapter.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Types\Comparer.cs" />
@@ -124,9 +125,7 @@
       <Private>False</Private>
     </ProjectReference>
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Execute\" />
-  </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
